Fix: Member 수정

See merge request s09-webmobile1-sub2/S09P12A307!52

+++ backend/src/main/java/signiel/heartsigniel/model/member/Member.java
@@ -44,7 +44,7 @@ public class Member {
     private boolean isBlack;
 
     @Column(name = "`rating`")
-    private Long rating;
+    private int rating;
 
     @Column(name = "`meeting_cnt`")
     private int meetingCount;
